Apply API review fixes for MTROTAHeaderParser.
* Move the parsing method to MTROTAHeader.
* Rename the file to MTROTAHeader.h
* Add documentation.
* Make the properties that can be missing nullable.
* Make properties copy instead of strong.

Fixes https://github.com/project-chip/connectedhomeip/issues/22540

Addresses part of https://github.com/project-chip/connectedhomeip/issues/22420
