+++ src/darwin/Framework/CHIP/MTROTAHeader.mm
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-#import "MTROTAHeaderParser.h"
+#import "MTROTAHeader.h"
 
 #import "MTRError.h"
 #import "MTRError_Internal.h"
@@ -25,9 +25,6 @@
 #include <lib/core/OTAImageHeader.h>
 
 @implementation MTROTAHeader
-@end
-
-@implementation MTROTAHeaderParser
 + (MTROTAHeader * _Nullable)headerFromData:(NSData *)data error:(NSError * __autoreleasing *)error
 {
     chip::OTAImageHeaderParser parser;
@@ -35,7 +32,9 @@
     parser.Init();
 
     if (!parser.IsInitialized()) {
-        *error = [NSError errorWithDomain:MTRErrorDomain code:MTRErrorCodeGeneralError userInfo:nil];
+        if (error != nil) {
+            *error = [NSError errorWithDomain:MTRErrorDomain code:MTRErrorCodeGeneralError userInfo:nil];
+        }
         return nil;
     }
 
@@ -43,7 +42,9 @@
     chip::OTAImageHeader header;
     CHIP_ERROR err = parser.AccumulateAndDecode(buffer, header);
     if (err != CHIP_NO_ERROR) {
-        *error = [MTRError errorForCHIPErrorCode:err];
+        if (error != nil) {
+            *error = [MTRError errorForCHIPErrorCode:err];
+        }
         parser.Clear();
         return nil;
     }
